requirements.txt
# Core News Processing
feedparser>=6.0.0
newspaper3k>=0.2.8
pyyaml>=6.0
edge-tts>=6.1.0
lxml>=4.9.0
tqdm>=4.64.0

# Multi-Provider LLM Support
python-dotenv>=1.0.0
openai>=1.0.0
requests>=2.28.0
google-generativeai>=0.3.0
ollama>=0.1.0

# Web Interface
flask>=2.3.0
werkzeug>=2.3.0
jinja2>=3.1.0
lxml_html_clean

# Optional: AI/ML (install separately if needed)
# torch>=2.0.0
# transformers>=4.30.0